Split out def & expand from `toc-hir`
