+++ compiler/toc-hir-def/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "toc-hir"
+name = "toc-hir-def"
 version = "0.1.0"
 
 authors.workspace = true
@@ -10,11 +10,14 @@ edition.workspace = true
 
 [dependencies]
 indexmap = { workspace = true }
-internment = "0.7.0"
 la-arena = "0.3"
-lazy_static = "1.4.0"
-toc-span = { workspace = true }
 toc-source-graph = { workspace = true }
+toc-syntax = { workspace = true }
+toc-vfs-db = { workspace = true }
+toc-ast-db = { workspace = true }
+rustc-hash = "1.1"
 
 toc-hir-expand.workspace = true
-toc-hir-def.workspace = true
+
+salsa = { workspace = true }
+upcast = { workspace = true }
